separate tests into unit & integration tests
