+++ test/integration/cli/.gitignore
