+++ test/integration/cli/entry.test.js
