+++ test/integration/cli/include.test.js
