+++ test/integration/configCases/InjectChangedModulesPlugin/devtool-cheap-module-source-map/index.js
