+++ test/integration/configCases/InjectChangedModulesPlugin/devtool-cheap-source-map/index.js
