+++ test/integration/configCases/InjectChangedModulesPlugin/devtool-module-source-map/index.js
