+++ test/integration/configCases/InjectChangedModulesPlugin/devtool-source-map/index.js
