+++ test/unit/cli/fixture/config/expected.json
