+++ test/unit/cli/fixture/webpackConfig/expected.json
