+++ test/unit/cli/fixture/webpackConfig/webpack.config-test.babel.js
