+++ test/unit/cli/fixture/webpackConfig/webpack.config-test.coffee
